updated based on comments
